second solution to Kata.java. Explanation of steps below:
line 5: instantiate a list to hold our digits.

line 8: use % to get the remainder of 'n' when divided by 10, which
is the last digit of the number, and add it to our list.

line 9: make use of the fact that longs cannot have decimal places and
divide by 10 to effectively chop off the last digit. If n is still
greater than 0 (line 7), repeat the process.

line 12 (aka the doosie): We need to convert our Integer list into an
int array. We take each element on the list (.stream()), map it from an
Integer to an int that is equal to the intValue of the Integer
(mapToInt(Integer::intValue)), then take the result, which is a list of
ints, and convert it to an array (.toArray()).

hatchi matchi